Nerf many monster spells
Because I suck at the game. :-P Joking aside, the only way to "play" the
game on deep dungeon levels was to dig anti-summoning corridors and use
LOS-blocking spells all the time, which doesn't really count as
"gameplay" for me. If there is only a single way to "play" the game, it
turns into the triple B - boring beyond belief!