Improvements
Pits should generate with a few items, like in newer versions of vanilla Angband. The princess bitch should have a valuable reward, which she does now in the form of one extra skill point whenever you save one. And the autoscum code had to be adjusted for the changed level feelings.