DUNGEON RANDOMIZER :D
Man, I can't believe that it works, but it works!!! Now, we can run this utility to randomize all the dungeons in the game (except god temples for now), just like randomizers for other games like Pokemon Emerald or Super Mario 64 :D It's wholly possible for the Barrow-Downs to be replaced with Mordor, Halls of Mandos or even the Void now!

For best results, this should be run before starting a new character and then the dungeon definition file be kept until that char dies, although I guess one could also "cheat" by re-randomizing it (or rolling back to the original dungeon definitions) during a running game.