Random quests early on
It can result in hugely out of depth monsters on dlvl1 if they can be 10 levels out of depth! So it's been made a little bit more sane now. Yes, I'm aware that this can result in stupidly easy quests on the first couple dlvls but then again the game isn't supposed to be super-mega-insta-deadly on the first couple dlvls, so there.