Fix macros for powers
Yeah I know, this implementation sucks, but then again the original one sucked a hundred times more. There's only two ways in which this code path can be called, anyway: either because you accidentally mistyped and used the U command with a char who doesn't have powers, in which case you just spam the spacebar, or because you have a macro that would activate a certain power and accidentally hit the macro's hotkey with a char who has no powers. And in the latter case, the results can be catastrophic if these extra --More--'s aren't in place!

***In a well-designed game you wouldn't need macros at all, see: Nethack***