New skill: dual-wielding
This one's being added to finally balance the total OPness of the Watcher in the Water. I've not had it myself, but read enough stories about it and know enough about the game that I can see how OP it is. Wielding three super-powerful artifacts with a character who gets 12 blows per round would mean you'd get *36 blows* per round; it's not hard to predict that nothing stands a chance against that. So there, now you'd need considerable amounts of dual-wielding skill and even then you still have a risk of not getting all the attacks.