FOOK YOO whoever created the deatmold race!!!
Like, how sloppy can you get, vomiting that race into the game without properly caring for it? And I said it before, just saying "well this is an experimental race" is an excuse that I'm not gonna accept! If I, a hobby programmer who doesn't even know C, can fix the race with a couple dozen commits, then anyone else could have done so as well and the fact that they didn't shows they're just *lazy*!