Message when geno doesn't work
Like wtf (again), why are you programmer mofos leaving the player clueless as to why on earth their genocide didn't work??? What's wrong with giving them a little message? You've made it so that there is a message when trying to geno in the death fate, too, so all you had to do was to *copy the damn message over*! That's something a five-year-old could do!