4D pocket skill
This was very hard to get to work. But now, the player can invest in a skill that allows the use of an extra storage because the damn inventory is way too small! For balance reasons, the amount of items you can store is limited by skill level (max 100) and taking stuff out works only while in a town (so that you can't simply use the pocket to protect consumables from destruction).