Ego item changes
WTF someone made it so that boomerangs cannot be blessed even though apparently it used to be possible! Also double WTF why is the elven DSM from that one quest unique??? DSM sucks in this game, might as well make it possible to spawn ones that suck a bit less!