Joke monsters cannot be disabled
And I hate hate hate the guts of the person who unconditionally disabled the Cth/Z monsters for no reason. The option was just fine, players could decide for themselves whether or not they want those monsters! To the person who took that option out: Would you like it if I disabled your option to engage in sexual activity IRL?