Artifact preservation versus Eru prayers
Now, that bug-ridden Eru prayer should no longer identify artifacts on the ground, which used to render them "known by the player". With this change, we can once again make it so that IDed artifacts on the ground count as created; this is important for artifacts like Travak which can spawn in multiples, otherwise you could just drop one and then find another stack to add to the one you already have!