Nerf summoning
Don't complain. Everyone and their mother knows that summoning was THE number one OP skill in ToME, so of course I had to do *some*thing. You can still use a partial totem of a GWOP and let it summon a hundred dragons, then dismiss the GWOP to stop the mana usage while keeping the 200 summoned pets for no upkeep!