Game remembers max recall depths
SO stupid to reset the "deepest level you reached" in a dungeon just because you decided to reset your recall depth to a lower value! That's brain-dead programming, if a dungeon has multiple towns this makes it needlessly hard to get to a deeper one after you've been to the shallower one!