Pattern levelport change
HOLY SHIT, you were getting a WIZMODE levelport for completing a pattern vault, even outside of wizmode? And it worked in dungeons that restrict levelporting??? Maybe that's why they took out that vault, because it'd have been SOOOOOOOO incredibly difficult to just change it to a random levelport to prevent skipping past Angband/Void/Nether Realm by scumming until a pattern vault spawns. Man, the abysses I discover when working with this game's mess of a source code... there's probably a hundred similar exploits still present which I'm not aware of...