Stacking rule change
It STILL doesn't fix the bug where "proofed" scrolls or potions sometimes appear in shops, displaying the egotype but not actually being proofed! What the hell, game! The egotype has 100% chance of giving the proofing flag, so why on earth do ego items spawn which lack the proofed flag??? That should be impossible but it keeps happening!