Display of finalized rods
It's SOOOOOOOOOOOOO endlessly stupid to display neither the amount of mana it costs per zap nor the description of the rod tip you used, because both of those can easily be accessed and displayed, plus you'd "know" what they are since you constructed the rod in the first place! Don't withhold that information from the player, goddammit!