Fix gigabug with quests
Holy shit, what a loophole. Why on earth does it restart those hooks when you save and load??? Thanks davidk64fnq for finding and especially reporting this error (as opposed to just keeping silent about it and abusing the hell out of it).

Vanilla ToME has this bug, too. You can play a Yavanna worshipper, go to the Mirkwood Spiders quest with a pack of scrolls/wands of summon monster, and then always save the game after completing the quest, load, and summon some monster. Due to the hooks having been improperly restarted, it'll count as "completing" the quest again and gives you another huge piety boost. I'd be surprised if any other ToME variants fixed that bug.