Nether realm fixes
Found out that it's as easy as placing an extra pink stair on the bottom level to avoid the fucked up march back UP 30 more levels of that crap. No other dungeon forces you to go through it twice, so the nether realm really shouldn't, either. Also players have reported that in vanilla ToME, towns on the nether realm have in the past caused them to get stuck due to the upstairs being bugged away on the way up, so this will allow you to circumvent such a fatal bug and finish your game!