Add even more traps
And capped the ridiculous "increase both damage dice and size" scaling factor of ball/bolt traps, plus fixed a fucking CRASH BUG in the polymorph code that the original ToME devs apparently didn't catch! Is that because polymorph is rare enough in a normal game so it normally doesn't come up?