Fix GIGABUG with special levels
Hooooooly shit, how on earth didn't anyone EVER spot that one in all those years when the ToME 2 series was a popular game??? Took me a while to get to the root of it but apparently the autoscummer was running even on special levels, despite those levels having a fixed layout and therefore absolutely zero reason to be subjected  to any kind of scumming! In particular, the "level feeling" of those should always be the same due to their non-randomness so I'm surprised the autoscummer didn't just try to reroll the level a thousand times, but oh well, at least it's finally fixed now.