FUCK
That's why you don't switch between different PCs when editing a file. Because otherwise you'll end up with two versions of the file, one of which has one half of your changes and the other file has the other half. At least in this case it was easy enough to fix though.